requirements.txt
aniso8601==1.2.0
appdirs==1.4.0
certifi==2019.11.28
chardet==3.0.4
click==6.7
Flask==0.12
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
idna==2.9
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
numpy==1.16.6
packaging==16.8
pandas==0.24.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.23.0
six==1.10.0
SQLAlchemy==1.1.5
urllib3==1.25.8
Werkzeug==0.11.15
